requirements.txt
python-dotenv==1.0.1
openai>=1.44.0
google-genai>=0.2.0
opencv-python==4.10.0.84
numpy==2.1.2
requests==2.32.3
moviepy>=1.0.3

